requirements.txt
flask
flask-socketio
flask-sqlalchemy
pyjwt
python-engineio
python-dotenv
eventlet
pymysql
cryptography